+++ sdk/crypto/crypto.hpp
@@ -213,7 +213,7 @@ class hash_t {
      * @param hash_context_t** handle [out]
      * @param hash_algorithm_t alg [in]
      * @param const unsigned char* key [inopt]
-     * @param unsigned keysize [inopt]
+     * @param size_t keysize [inopt]
      * @return error code (see error.hpp)
      * @example
      *    binary_t hash_data;
@@ -242,13 +242,13 @@ class hash_t {
      *    hash.hash(handle, source, source_size, hash_data);
      *    hash.close(handle)
      */
-    virtual return_t open(hash_context_t** handle, const char* algorithm, const unsigned char* key = nullptr, unsigned keysize = 0) = 0;
+    virtual return_t open(hash_context_t** handle, const char* algorithm, const unsigned char* key = nullptr, size_t keysize = 0) = 0;
     /**
      * @brief open (HMAC, CMAC)
      */
     virtual return_t open(hash_context_t** handle, const char* algorithm, const binary_t& key) = 0;
 
-    virtual return_t open(hash_context_t** handle, hash_algorithm_t alg, const unsigned char* key = nullptr, unsigned keysize = 0) = 0;
+    virtual return_t open(hash_context_t** handle, hash_algorithm_t alg, const unsigned char* key = nullptr, size_t keysize = 0) = 0;
     /**
      * @brief open (HMAC)
      */
@@ -259,10 +259,10 @@ class hash_t {
      * @param crypt_algorithm_t alg [in]
      * @param crypt_mode_t mode [in]
      * @param const unsigned char* key [in]
-     * @param unsigned keysize [in]
+     * @param size_t keysize [in]
      * @return error code (see error.hpp)
      */
-    virtual return_t open(hash_context_t** handle, crypt_algorithm_t alg, crypt_mode_t mode, const unsigned char* key, unsigned keysize) = 0;
+    virtual return_t open(hash_context_t** handle, crypt_algorithm_t alg, crypt_mode_t mode, const unsigned char* key, size_t keysize) = 0;
     /**
      * @brief open (CMAC)
      */
